- split SyncReadStream and AsyncWriteStream
